Quiz Page Now we have a quiz page, where all the public quizzes can be seen. Quiz deletion is also in place.
Simultaneously moving the previous Quiz page to a test page, that still needs to be figured out.
We'll need 2 pages, one for attempting a quiz and 2 for building the quiz data (question).

CRUD for Quiz Page Data
Completed the Crud Functionality for Quiz Page,
fixed the import path issue.
fixed quiz path issues.